+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 #Changelog
 
+## 2.0.0
+* `init()` removed from the AlgoliaComponent ![probably won't](https://img.shields.io/badge/will%20it%20affect%20me%3F-probably%20won't-green.svg)
+* `createManager()` visibility changed to private in AlgoliaComponent ![probably won't](https://img.shields.io/badge/will%20it%20affect%20me%3F-probably%20won't-green.svg)
+* `yii\base\InvalidConfigException` is thrown instead of `Exception` in case of invalid configuration for the AlgoliaComponent ![probably won't](https://img.shields.io/badge/will%20it%20affect%20me%3F-probably%20won't-green.svg)
+*
+
 ## 1.3.0
 * Updated the package to depend on ^1.16.0 versions of the [official Algolia client](https://github.com/algolia/algoliasearch-client-php)  
 * Removed methods marked as deprecated from the `leinonen\Yii2Algolia\AlgoliaManager` & `leinonen\Yii2Algolia\AlgoliaManager` @method phpdocs. The methods are still found in the `AlgoliaSearch\Client` so no existing code will break, just no IDE autocompletion anymore. Expect these to be removed in 2.0.0.  This includes methods: